# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto controle de Filmes - Senai
  description: |-
   API destinada ao controle de filmes para o projeto "BEM LEGAL" do SENAI
  termsOfService: http://swagger.io/terms/
  contact:
    email: seunome@email.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.5.26
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: Endopoint para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '415':
          description: Invalid Content - Type
        '500':
          description: Error Internal Sever (Controller / Model)
    get:
      tags:
        - Filme
      summary: Lista todos os filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Not Found
        '500':
          description: Error Internal Sever (Controller / Model)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - aluno
      summary: Busca um Filme pelo ID
      description: Retorna um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
        '500':
          description: Error Internal Sever (Controller / Model)
    put:
      tags:
        - Filme
      summary: Atualiza um filme pelo ID
      description: Buscar e atualizar um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID ID para realizar a busca do Filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um Filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found 
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internal Sever (Controller / Model)
    delete:
      tags:
        - Filme
      summary: Excluir um Filme pelo id
      description: buscar e Excluir um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para excluir um filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
        '500': 
          description: Error Internal Sever (Controller / Model)
components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
        - data_lancamento
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
        duracao:
          type: string
        valor:
          type: integer
        avaliacao:
          type: integer
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    